+++ GoldenFeets/src/main/java/com/GoldenFeet/GoldenFeets/repository/CategoriaRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Long> { // <-- CORRECCIÓN AQUÍ
-    // No se necesitan métodos personalizados por ahora.
+
+public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
+
 }
